requirements.txt
amqp==5.1.1
ansicon==1.89.0
arabic-reshaper==3.0.0
arrow==1.2.3
asgiref==3.6.0
asn1crypto==1.5.1
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
blessed==1.20.0
boto3==1.26.146
botocore==1.29.146
Brotli==1.0.9
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
crispy-bootstrap5==0.7
cron-descriptor==1.4.5
cryptography==41.0.1
cssselect2==0.7.0
Django==4.2.1
django-appconf==1.0.5
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-compat==1.0.15
django-compressor==4.3.1
django-crispy-forms==2.0
django-crontab==0.7.1
django-dynamic-preferences==1.15.0
django-glrm==1.1.3
django-mailgun-mime==0.1.7
django-mjml==1.3
django-picklefield==3.1
django-q==1.3.9
django-session-timeout==0.1.0
django-timezone-field==7.0
exceptiongroup==1.1.3
fontawesomefree==6.4.0
fonttools==4.40.0
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.4
jinxed==1.2.0
jmespath==1.0.1
kombu==5.3.0
lxml==4.9.3
oscrypto==1.3.0
outcome==1.2.0
packaging==24.2
persisting-theory==1.0
Pillow==9.5.0
pip-system-certs==4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycparser==2.21
pydyf==0.7.0
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
pypdf==3.12.0
pyphen==0.14.0
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-crontab==3.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
rcssmin==1.1.1
redis==3.5.3
reportlab==3.6.13
requests==2.31.0
rjsmin==1.2.1
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.16
uvicorn==0.34.0
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
wrapt==1.16.0
wsproto==1.2.0
xhtml2pdf==0.2.11
xlwt==1.3.0
zopfli==0.2.2
django-qr-code